FeLiNa

+++ Controllers/NotificationsController.cs
@@ -143,9 +143,9 @@ namespace Biz_collab.Controllers
             {
                 return NotFound();
             }
-            if (_db.MutedNames.Any(m => m.Name == name))
+            if (_db.MutedNames.Any(m => m.ClientId== this.User.FindFirst(ClaimTypes.NameIdentifier).Value && m.Name == name))
             {
-                var n = _db.MutedNames.FirstOrDefault(m => m.Name == name);
+                var n = _db.MutedNames.FirstOrDefault(m => m.ClientId == this.User.FindFirst(ClaimTypes.NameIdentifier).Value && m.Name == name);
                 _db.MutedNames.Remove(n);
             }
             else {
@@ -166,7 +166,8 @@ namespace Biz_collab.Controllers
             {
                 return NotFound();
             }
-            if (act) {// удалить и отказаться от приглашения(отослать уведомление об отказе тому кто отправил приглашение)
+            if (act && !_db.MutedNames.Any(m => m.ClientId == _db.Clients.FirstOrDefault(c => c.Login == login).Id && m.Name == this.User.FindFirst(ClaimTypes.Name).Value)) {// удалить и отказаться от приглашения(отослать уведомление об отказе тому кто отправил приглашение)
+                
                 Notification declineOffer = new Notification
                 {
                     ClientId =  _db.Clients.FirstOrDefault(c=>c.Login==login).Id,
